Revert "[RHCLOUD-40718] Cleanup Notifications exposed apis (#3514)"
This reverts commit 6f318e675d40812ab5f6f0231356fe14ae99b3f2.
