+++ backend/src/main/java/com/redhat/cloud/notifications/routers/handlers/drawer/DrawerResource.java
@@ -40,11 +40,8 @@ import static com.redhat.cloud.notifications.routers.SecurityContextUtil.getOrgI
 import static com.redhat.cloud.notifications.routers.SecurityContextUtil.getUsername;
 import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON;
 
+@Path(API_NOTIFICATIONS_V_1_0 + "/notifications/drawer")
 public class DrawerResource {
-
-    @Path(API_NOTIFICATIONS_V_1_0 + "/notifications/drawer")
-    public static class V1 extends DrawerResource {
-    }
 
     @Inject
     DrawerNotificationRepository drawerRepository;
