+++ backend/src/main/java/com/redhat/cloud/notifications/routers/handlers/endpoint/EndpointResource.java
@@ -1,9 +1,11 @@
 package com.redhat.cloud.notifications.routers.handlers.endpoint;
 
+import com.redhat.cloud.notifications.Constants;
 import com.redhat.cloud.notifications.auth.ConsoleIdentityProvider;
 import com.redhat.cloud.notifications.auth.annotation.Authorization;
 import com.redhat.cloud.notifications.auth.annotation.IntegrationId;
 import com.redhat.cloud.notifications.auth.kessel.KesselAssets;
+import com.redhat.cloud.notifications.auth.kessel.KesselAuthorization;
 import com.redhat.cloud.notifications.auth.kessel.KesselInventoryAuthorization;
 import com.redhat.cloud.notifications.auth.kessel.permission.IntegrationPermission;
 import com.redhat.cloud.notifications.auth.kessel.permission.WorkspacePermission;
@@ -17,16 +19,22 @@ import com.redhat.cloud.notifications.db.repositories.EndpointEventTypeRepositor
 import com.redhat.cloud.notifications.db.repositories.EndpointRepository;
 import com.redhat.cloud.notifications.db.repositories.EventTypeRepository;
 import com.redhat.cloud.notifications.db.repositories.NotificationRepository;
+import com.redhat.cloud.notifications.models.Application;
 import com.redhat.cloud.notifications.models.BehaviorGroup;
 import com.redhat.cloud.notifications.models.Bundle;
 import com.redhat.cloud.notifications.models.CamelProperties;
+import com.redhat.cloud.notifications.models.CompositeEndpointType;
 import com.redhat.cloud.notifications.models.Endpoint;
 import com.redhat.cloud.notifications.models.EndpointProperties;
 import com.redhat.cloud.notifications.models.EndpointStatus;
 import com.redhat.cloud.notifications.models.EndpointType;
+import com.redhat.cloud.notifications.models.EventType;
 import com.redhat.cloud.notifications.models.SourcesSecretable;
 import com.redhat.cloud.notifications.models.SystemSubscriptionProperties;
+import com.redhat.cloud.notifications.models.dto.v1.ApplicationDTO;
+import com.redhat.cloud.notifications.models.dto.v1.BundleDTO;
 import com.redhat.cloud.notifications.models.dto.v1.CommonMapper;
+import com.redhat.cloud.notifications.models.dto.v1.EventTypeDTO;
 import com.redhat.cloud.notifications.models.dto.v1.NotificationHistoryDTO;
 import com.redhat.cloud.notifications.models.dto.v1.endpoint.EndpointDTO;
 import com.redhat.cloud.notifications.models.dto.v1.endpoint.EndpointMapper;
@@ -76,15 +84,17 @@ import org.jboss.resteasy.reactive.RestPath;
 
 import java.net.URI;
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
-import static com.redhat.cloud.notifications.Constants.API_INTEGRATIONS_V_1_0;
 import static com.redhat.cloud.notifications.db.Query.DEFAULT_RESULTS_PER_PAGE;
 import static com.redhat.cloud.notifications.models.EndpointType.CAMEL;
 import static com.redhat.cloud.notifications.models.EndpointType.DRAWER;
@@ -97,11 +107,12 @@ import static jakarta.ws.rs.core.MediaType.TEXT_PLAIN;
 
 // Email endpoints are not added at this point
 // TODO Needs documentation annotations
-public class EndpointResource extends EndpointResourceCommon {
+public class EndpointResource {
 
     public static final String DEPRECATED_SLACK_CHANNEL_ERROR = "The channel field is deprecated";
     public static final String HTTPS_ENDPOINT_SCHEME_REQUIRED = "The endpoint URL must start with \"https\"";
     public static final String UNSUPPORTED_ENDPOINT_TYPE = "Unsupported endpoint type";
+    public static final String REDACTED_CREDENTIAL = "*****";
     public static final String AUTO_CREATED_BEHAVIOR_GROUP_NAME_TEMPLATE = "Integration \"%s\" behavior group";
 
     @Inject
@@ -119,6 +130,9 @@ public class EndpointResource extends EndpointResourceCommon {
 
     @Inject
     KesselAssets kesselAssets;
+
+    @Inject
+    KesselAuthorization kesselAuthorization;
 
     @Inject
     KesselInventoryAuthorization kesselInventoryAuthorization;
@@ -150,42 +164,42 @@ public class EndpointResource extends EndpointResourceCommon {
     @Inject
     SecretUtils secretUtils;
 
-    @Path(API_INTEGRATIONS_V_1_0 + "/endpoints")
+    @Path(Constants.API_INTEGRATIONS_V_1_0 + "/endpoints")
     static class V1 extends EndpointResource {
-    }
 
-    @GET
-    @Path("/{id}/history")
-    @Produces(APPLICATION_JSON)
-    @Parameters({
-        @Parameter(
-            name = "limit",
-            in = ParameterIn.QUERY,
-            description = "Number of items per page, if not specified " + DEFAULT_RESULTS_PER_PAGE + " is used.",
-            schema = @Schema(type = SchemaType.INTEGER, defaultValue = DEFAULT_RESULTS_PER_PAGE + "")
-            ),
-        @Parameter(
-            name = "pageNumber",
-            in = ParameterIn.QUERY,
-            description = "Page number. Starts at first page (0), if not specified starts at first page.",
-            schema = @Schema(type = SchemaType.INTEGER)
-            ),
-        @Parameter(
-            name = "includeDetail",
-            description = "Include the detail in the reply",
-            schema = @Schema(type = SchemaType.BOOLEAN)
-            )
-    })
-    @Authorization(legacyRBACRole = ConsoleIdentityProvider.RBAC_READ_INTEGRATIONS_ENDPOINTS, integrationPermissions = {IntegrationPermission.VIEW_HISTORY})
-    public List<NotificationHistoryDTO> getEndpointHistory(@Context SecurityContext sec, @IntegrationId @PathParam("id") UUID id, @QueryParam("includeDetail") Boolean includeDetail, @Valid @BeanParam Query query) {
-        if (!this.endpointRepository.existsByUuidAndOrgId(id, getOrgId(sec))) {
-            throw new NotFoundException("Endpoint not found");
+        @GET
+        @Path("/{id}/history")
+        @Produces(APPLICATION_JSON)
+        @Parameters({
+            @Parameter(
+                name = "limit",
+                in = ParameterIn.QUERY,
+                description = "Number of items per page, if not specified " + DEFAULT_RESULTS_PER_PAGE + " is used.",
+                schema = @Schema(type = SchemaType.INTEGER, defaultValue = DEFAULT_RESULTS_PER_PAGE + "")
+                ),
+            @Parameter(
+                name = "pageNumber",
+                in = ParameterIn.QUERY,
+                description = "Page number. Starts at first page (0), if not specified starts at first page.",
+                schema = @Schema(type = SchemaType.INTEGER)
+                ),
+            @Parameter(
+                name = "includeDetail",
+                description = "Include the detail in the reply",
+                schema = @Schema(type = SchemaType.BOOLEAN)
+                )
+        })
+        @Authorization(legacyRBACRole = ConsoleIdentityProvider.RBAC_READ_INTEGRATIONS_ENDPOINTS, integrationPermissions = {IntegrationPermission.VIEW_HISTORY})
+        public List<NotificationHistoryDTO> getEndpointHistory(@Context SecurityContext sec, @IntegrationId @PathParam("id") UUID id, @QueryParam("includeDetail") Boolean includeDetail, @Valid @BeanParam Query query) {
+            if (!this.endpointRepository.existsByUuidAndOrgId(id, getOrgId(sec))) {
+                throw new NotFoundException("Endpoint not found");
+            }
+
+            // TODO We need globally limitations (Paging support and limits etc)
+            String orgId = getOrgId(sec);
+            boolean doDetail = includeDetail != null && includeDetail;
+            return commonMapper.notificationHistoryListToNotificationHistoryDTOList(notificationRepository.getNotificationHistory(orgId, id, doDetail, query));
         }
-
-        // TODO We need globally limitations (Paging support and limits etc)
-        String orgId = getOrgId(sec);
-        boolean doDetail = includeDetail != null && includeDetail;
-        return commonMapper.notificationHistoryListToNotificationHistoryDTOList(notificationRepository.getNotificationHistory(orgId, id, doDetail, query));
     }
 
     @GET
@@ -235,6 +249,67 @@ public class EndpointResource extends EndpointResourceCommon {
         }
 
         return internalGetEndpoints(sec, query, targetType, activeOnly, name, authorizedIds, false);
+    }
+
+    /**
+     * Gets the list of endpoints.
+     * @param sec the security context of the request.
+     * @param query the page related query elements.
+     * @param targetType the types of the endpoints to fetch.
+     * @param activeOnly should only the active endpoints be fetched?
+     * @param name filter endpoints by name.
+     * @param authorizedIds set of authorized integrations that we are allowed
+     *                      to fetch.
+     * @return a page containing the requested endpoints.
+     */
+    protected EndpointPage internalGetEndpoints(
+        final SecurityContext sec,
+        final Query query,
+        final List<String> targetType,
+        final Boolean activeOnly,
+        final String name,
+        final Set<UUID> authorizedIds,
+        final boolean includeLinkedEventTypes
+    ) {
+        String orgId = getOrgId(sec);
+
+        List<Endpoint> endpoints;
+        Long count;
+
+        Set<CompositeEndpointType> compositeType;
+
+        if (targetType != null && targetType.size() > 0) {
+            compositeType = targetType.stream().map(s -> {
+                try {
+                    return CompositeEndpointType.fromString(s);
+                } catch (IllegalArgumentException e) {
+                    throw new BadRequestException("Unknown endpoint type: [" + s + "]", e);
+                }
+            }).collect(Collectors.toSet());
+        } else {
+            compositeType = Set.of();
+        }
+
+        endpoints = endpointRepository.getEndpointsPerCompositeType(orgId, name, compositeType, activeOnly, query, authorizedIds);
+        count = endpointRepository.getEndpointsCountPerCompositeType(orgId, name, compositeType, activeOnly, authorizedIds);
+
+        final List<EndpointDTO> endpointDTOS = new ArrayList<>(endpoints.size());
+        for (Endpoint endpoint: endpoints) {
+            // Fetch the secrets from Sources.
+            this.secretUtils.loadSecretsForEndpoint(endpoint);
+
+            // Redact the secrets for the endpoint if the user does not have
+            // permission.
+            this.redactSecretsForEndpoint(sec, endpoint);
+
+            EndpointDTO endpointDTO = endpointMapper.toDTO(endpoint);
+            if (includeLinkedEventTypes) {
+                includeLinkedEventTypes(endpoint.getEventTypes(), endpointDTO);
+            }
+            endpointDTOS.add(endpointDTO);
+        }
+
+        return new EndpointPage(endpointDTOS, new HashMap<>(), new Meta(count));
     }
 
     @POST
@@ -426,6 +501,26 @@ public class EndpointResource extends EndpointResourceCommon {
     @Authorization(legacyRBACRole = ConsoleIdentityProvider.RBAC_READ_INTEGRATIONS_ENDPOINTS, integrationPermissions = {IntegrationPermission.VIEW})
     public EndpointDTO getEndpoint(@Context SecurityContext sec, @IntegrationId @PathParam("id") UUID id) {
         return internalGetEndpoint(sec, id, false);
+    }
+
+    protected EndpointDTO internalGetEndpoint(final SecurityContext securityContext, final UUID id, final boolean includeLinkedEventTypes) {
+        String orgId = getOrgId(securityContext);
+        Optional<Endpoint> endpoint = endpointRepository.getEndpointWithLinkedEventTypes(orgId, id);
+        if (endpoint.isEmpty()) {
+            throw new NotFoundException();
+        } else {
+            // Fetch the secrets from Sources.
+            this.secretUtils.loadSecretsForEndpoint(endpoint.get());
+
+            // Redact all the credentials from the endpoint's properties.
+            this.redactSecretsForEndpoint(securityContext, endpoint.get());
+
+            EndpointDTO endpointDTO = this.endpointMapper.toDTO(endpoint.get());
+            if (includeLinkedEventTypes) {
+                includeLinkedEventTypes(endpoint.get().getEventTypes(), endpointDTO);
+            }
+            return endpointDTO;
+        }
     }
 
     @DELETE
@@ -657,6 +752,50 @@ public class EndpointResource extends EndpointResourceCommon {
         return !backendConfig.isEmailsOnlyModeEnabled() || endpointType.isSystemEndpointType;
     }
 
+    /**
+     * Removes the secrets from the endpoint's properties when returning them
+     * to the client.
+     * @param endpoint the endpoint to redact the secrets from.
+     */
+    @Deprecated(forRemoval = true)
+    protected void redactSecretsForEndpoint(final SecurityContext securityContext, final Endpoint endpoint) {
+        // Figure out if the principal has "write" permissions on the
+        // integration or not, to decide whether we should redact the secrets
+        // from the returning payload.
+        //
+        // Users with just read permissions will get the secrets redacted for
+        // them.
+        boolean shouldRedactSecrets;
+        if (this.backendConfig.isKesselRelationsEnabled(getOrgId(securityContext))) {
+            try {
+                if (this.backendConfig.isKesselInventoryUseForPermissionsChecksEnabled(getOrgId(securityContext))) {
+                    this.kesselInventoryAuthorization.hasPermissionOnIntegration(securityContext, IntegrationPermission.EDIT, endpoint.getId());
+                } else {
+                    this.kesselAuthorization.hasPermissionOnIntegration(securityContext, IntegrationPermission.EDIT, endpoint.getId());
+                }
+                shouldRedactSecrets = false;
+            } catch (final ForbiddenException | NotFoundException e) {
+                shouldRedactSecrets = true;
+            }
+        } else {
+            shouldRedactSecrets = !securityContext.isUserInRole(ConsoleIdentityProvider.RBAC_WRITE_INTEGRATIONS_ENDPOINTS);
+        }
+
+        if (shouldRedactSecrets) {
+            if (endpoint.getProperties() instanceof SourcesSecretable sourcesSecretable) {
+                final String bearerToken = sourcesSecretable.getBearerAuthentication();
+                if (bearerToken != null) {
+                    sourcesSecretable.setBearerAuthentication(REDACTED_CREDENTIAL);
+                }
+
+                final String secretToken = sourcesSecretable.getSecretToken();
+                if (secretToken != null) {
+                    sourcesSecretable.setSecretToken(REDACTED_CREDENTIAL);
+                }
+            }
+        }
+    }
+
     @DELETE
     @Path("/{endpointId}/eventType/{eventTypeId}")
     @Operation(summary = "Delete the link between an endpoint and an event type", description = "Delete the link between an endpoint and an event type.")
@@ -786,4 +925,33 @@ public class EndpointResource extends EndpointResourceCommon {
         }
     }
 
+    private void includeLinkedEventTypes(Set<EventType> eventTypes, EndpointDTO endpointDTO) {
+        if (null != eventTypes && !eventTypes.isEmpty()) {
+            Map<Application, List<EventType>> applicationMap = eventTypes.stream()
+                .sorted(Comparator.comparing(EventType::getDisplayName))
+                .collect(Collectors.groupingBy(EventType::getApplication));
+            Map<Bundle, List<Application>> bundleMap = applicationMap.keySet().stream()
+                .sorted(Comparator.comparing(Application::getDisplayName))
+                .collect(Collectors.groupingBy(Application::getBundle));
+
+            List<Bundle> bundleList = bundleMap.keySet().stream().sorted(Comparator.comparing(Bundle::getDisplayName)).toList();
+
+            Set<BundleDTO> bundleDTOSet = new LinkedHashSet<>();
+            for (Bundle bundle : bundleList) {
+                Set<ApplicationDTO> applicationDTOSet = new LinkedHashSet<>();
+                List<Application> applications = bundleMap.get(bundle);
+                for (Application application : applications) {
+                    ApplicationDTO applicationDTO = commonMapper.applicationToApplicationDTO(application);
+                    Set<EventTypeDTO> eventTypesDTO = new LinkedHashSet<>();
+                    eventTypesDTO.addAll(commonMapper.eventTypeListToEventTypeDTOList(applicationMap.get(application)));
+                    applicationDTO.setEventTypes(eventTypesDTO);
+                    applicationDTOSet.add(applicationDTO);
+                }
+                BundleDTO bundleDTO = commonMapper.bundleToBundleDTO(bundle);
+                bundleDTO.setApplications(applicationDTOSet);
+                bundleDTOSet.add(bundleDTO);
+            }
+            endpointDTO.setEventTypesGroupByBundlesAndApplications(bundleDTOSet);
+        }
+    }
 }
