+++ backend/src/main/java/com/redhat/cloud/notifications/routers/handlers/event/EventResource.java
@@ -55,11 +55,8 @@ import static com.redhat.cloud.notifications.db.Query.DEFAULT_RESULTS_PER_PAGE;
 import static com.redhat.cloud.notifications.routers.SecurityContextUtil.getOrgId;
 import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON;
 
+@Path(API_NOTIFICATIONS_V_1_0 + "/notifications/events")
 public class EventResource {
-
-    @Path(API_NOTIFICATIONS_V_1_0 + "/notifications/events")
-    public static class V1 extends EventResource {
-    }
 
     public static final String TOTAL_RECIPIENTS = "total_recipients";
