+++ backend/src/test/java/com/redhat/cloud/notifications/openapi/OpenApiTest.java
@@ -2,7 +2,6 @@ package com.redhat.cloud.notifications.openapi;
 
 import com.redhat.cloud.notifications.oapi.OApiFilter;
 import com.reprezen.kaizen.oasparser.OpenApi3Parser;
-import com.reprezen.kaizen.oasparser.OpenApiParser;
 import com.reprezen.kaizen.oasparser.model3.OpenApi3;
 import com.reprezen.kaizen.oasparser.val.ValidationResults;
 import io.quarkus.logging.Log;
@@ -12,7 +11,6 @@ import io.swagger.parser.OpenAPIParser;
 import io.swagger.v3.parser.core.models.SwaggerParseResult;
 import org.junit.jupiter.api.Test;
 
-import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
 import java.nio.file.Files;
@@ -20,7 +18,6 @@ import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import static io.restassured.RestAssured.given;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -46,29 +43,12 @@ public class OpenApiTest {
     @TestHTTPResource("/api/doesNotExist/v1.0/openapi.json")
     URL badUrl;
 
-    private Optional<OpenApi3> getModelFromUrl(URL url) throws Exception {
-        try {
-            OpenApi3 model = new OpenApi3Parser().parse(url, true);
-            return Optional.of(model);
-        } catch (OpenApiParser.OpenApiParserException e) {
-            if (!url.toString().contains("/private/v2")) {
-                throw e;
-            }
-        }
-        return Optional.empty();
-    }
-
     @Test
     public void validateOpenApi() throws Exception {
         List<URL> urls = buildApiUrls(true);
 
         for (URL url : urls) {
-            final Optional<OpenApi3> optModel = getModelFromUrl(url);
-            if (optModel.isEmpty()) {
-                continue;
-            }
-            final OpenApi3 model = optModel.get();
-
+            OpenApi3 model = new OpenApi3Parser().parse(url, true);
             Log.infof("testing api %s", url);
             if (!model.isValid()) {
                 for (ValidationResults.ValidationItem item : model.getValidationItems()) {
@@ -96,23 +76,6 @@ public class OpenApiTest {
     }
 
     @Test
-    void testPrivateV2DontExist() {
-        given()
-            .accept("application/json")
-            .when()
-            .get("/api/private/v2.0/openapi.json")
-            .then()
-            .statusCode(404);
-
-        given()
-            .accept("application/json")
-            .when()
-            .get("/api/private/v2/openapi.json")
-            .then()
-            .statusCode(404);
-    }
-
-    @Test
     void exportOpenApiFile() throws Exception {
 
         List<URL> urls = buildApiUrls(true);
@@ -122,11 +85,6 @@ public class OpenApiTest {
                 java.nio.file.Path path = Paths.get("./target/" + url.getPath());
                 Files.createDirectories(path.getParent());
                 Files.copy(in, path, StandardCopyOption.REPLACE_EXISTING);
-            } catch (IOException e) {
-                if (url.toString().contains("/private/v2")) {
-                    continue;
-                }
-                throw e;
             }
         }
     }
